requirements.txt
berserk==0.12.8
black==23.7.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
Deprecated==1.2.14
idna==3.4
mypy-extensions==1.0.0
ndjson==0.3.1
packaging==23.1
pathspec==0.11.1
platformdirs==3.9.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pyxel==1.9.18
requests==2.31.0
six==1.16.0
urllib3==2.0.4
wrapt==1.15.0
